requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
chardet==3.0.4
click==6.7
colorama==0.3.7
constantly==15.1.0
cryptography==2.1.4
deluge==1.3.15
Django==1.11.14
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2
enum34==1.1.6
folder-color==0.0.86
folder-color-common==0.0.86
gyp==0.1
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
Markdown==2.6.11
numpy==1.13.3
PAM==0.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pygame===1.9.1release
pygobject==3.26.1
pyOpenSSL==17.5.0
pyserial==3.4
pytz==2018.5
pyxdg==0.25
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
Twisted==17.9.0
zope.interface==4.3.2
